nukiPyBridge/requirements.txt
pynacl
crc16
pybluez
pexpect
pygatt
